[MOV] utm: organize module structure according to guidelines
This module holds old code and is still not updated according to guidelines.
Let us update a bit the naming in this module to have readable code.

This commit is linked to task ID 1896681 and PR #27975. No functional change
should occur with this commit as it contains only renaming / move.
