+++ addons/utm/__init__.py
@@ -1,2 +1,4 @@
 # -*- coding: utf-8 -*-
+# Part of Odoo. See LICENSE file for full copyright and licensing details.
+
 from . import models
