+++ addons/utm/__manifest__.py
@@ -1,16 +1,18 @@
+# -*- coding: utf-8 -*-
+# Part of Odoo. See LICENSE file for full copyright and licensing details.
+
 {
     'name': 'UTM Trackers',
     'category': 'Hidden',
     'description': """
-Enable UTM trackers in shared links.
-=====================================================
-        """,
+Enable management of UTM trackers: campaign, medium, source.
+""",
     'version': '1.0',
     'depends': ['base'],
     'data': [
+        'data/utm_data.xml',
+        'views/utm_views.xml',
         'security/ir.model.access.csv',
-        'views/utm.xml',
-        'data/utm_data.xml'
     ],
     'demo': [],
     'auto_install': False,
