+++ addons/utm/models/__init__.py
@@ -1,3 +1,6 @@
 # -*- coding: utf-8 -*-
+# Part of Odoo. See LICENSE file for full copyright and licensing details.
+
 from . import utm
+from . import utm_mixin
 from . import ir_http
